defunct some tests The payment tests was already implicitly performing the other tests
