+++ client_test/payment_with_p2p_ms_test.go
@@ -1,16 +1,31 @@
 package client_test
 
 import (
+	"io"
+	"math/big"
 	"testing"
 
+	"github.com/statechannels/go-nitro/client"
+	"github.com/statechannels/go-nitro/client/engine"
 	"github.com/statechannels/go-nitro/client/engine/chainservice"
 	p2pms "github.com/statechannels/go-nitro/client/engine/messageservice/p2p-message-service"
+	"github.com/statechannels/go-nitro/client/engine/store"
+	td "github.com/statechannels/go-nitro/internal/testdata"
+	"github.com/statechannels/go-nitro/protocols"
 )
 
-func TestVirtualFundWithP2PMessageService(t *testing.T) {
+// setupClientWithP2PMessageService is a helper function that contructs a client and returns the new client and its store.
+func setupClientWithP2PMessageService(pk []byte, port int, chain *chainservice.MockChainService, logDestination io.Writer) (client.Client, *p2pms.P2PMessageService) {
+
+	messageservice := p2pms.NewMessageService("127.0.0.1", port, pk)
+	storeA := store.NewMemStore(pk)
+	return client.New(messageservice, chain, storeA, logDestination, &engine.PermissivePolicy{}, nil), messageservice
+}
+
+func TestPayments(t *testing.T) {
 
 	// Setup logging
-	logFile := "test_virtual_fund_with_simple_tcp.log"
+	logFile := "test_payments.log"
 	truncateLog(logFile)
 	logDestination := newLogWriter(logFile)
 
@@ -22,7 +37,6 @@ func TestVirtualFundWithP2PMessageService(t *testing.T) {
 	clientA, msgA := setupClientWithP2PMessageService(alice.PrivateKey, 3005, chainServiceA, logDestination)
 	clientB, msgB := setupClientWithP2PMessageService(bob.PrivateKey, 3006, chainServiceB, logDestination)
 	clientI, msgI := setupClientWithP2PMessageService(irene.PrivateKey, 3007, chainServiceI, logDestination)
-
 	peers := []p2pms.PeerInfo{
 		{Id: msgA.Id(), IpAddress: "127.0.0.1", Port: 3005, Address: alice.Address()},
 		{Id: msgB.Id(), IpAddress: "127.0.0.1", Port: 3006, Address: bob.Address()},
@@ -36,11 +50,20 @@ func TestVirtualFundWithP2PMessageService(t *testing.T) {
 	defer msgA.Close()
 	defer msgB.Close()
 	defer msgI.Close()
+
 	directlyFundALedgerChannel(t, clientA, clientI)
 	directlyFundALedgerChannel(t, clientI, clientB)
+	outcome := td.Outcomes.Create(alice.Address(), bob.Address(), 100, 100)
+	r := clientA.CreateVirtualPaymentChannel(irene.Address(), bob.Address(), 0, outcome)
 
-	ids := createVirtualChannels(clientA, bob.Address(), irene.Address(), 5)
+	ids := []protocols.ObjectiveId{r.Id}
+
 	waitTimeForCompletedObjectiveIds(t, &clientA, defaultTimeout, ids...)
 	waitTimeForCompletedObjectiveIds(t, &clientB, defaultTimeout, ids...)
 	waitTimeForCompletedObjectiveIds(t, &clientI, defaultTimeout, ids...)
+	clientA.Pay(r.ChannelId, big.NewInt(5))
+
+	expected := BasicVoucherInfo{big.NewInt(5), r.ChannelId}
+	waitTimeForReceivedVoucher(t, &clientB, defaultTimeout, expected)
+
 }
